+++ src/ansys/fluent/core/services/transcript.py
@@ -1,46 +1,20 @@
 """Wrapper over the transcript grpc service of Fluent."""
 
-from typing import Generator, List, Tuple
+from typing import List, Tuple
 
 import grpc
 
 from ansys.api.fluent.v0 import transcript_pb2 as TranscriptModule
 from ansys.api.fluent.v0 import transcript_pb2_grpc as TranscriptGrpcModule
+from ansys.fluent.core.services.streaming import StreamingService
 
 
-class TranscriptService:
-    """Class wrapping the transcript gRPC service of Fluent.
-
-    Methods
-    -------
-    begin_streaming
-        Begin transcript streaming from Fluent.
-    """
+class TranscriptService(StreamingService):
+    """Class wrapping the transcript gRPC service of Fluent."""
 
     def __init__(self, channel: grpc.Channel, metadata: List[Tuple[str, str]]):
-        self.__stub = TranscriptGrpcModule.TranscriptStub(channel)
-        self.__metadata = metadata
-        self.__streams = None
-
-    def begin_streaming(
-        self, started_evt
-    ) -> Generator[TranscriptModule.TranscriptResponse, None, None]:
-        """Begin transcript streaming from Fluent.
-
-        Yields
-        ------
-        str
-            A transcript line
-        """
-        request = TranscriptModule.TranscriptRequest()
-        self.__streams = self.__stub.BeginStreaming(request, metadata=self.__metadata)
-        started_evt.set()
-        while True:
-            try:
-                yield next(self.__streams)
-            except Exception:
-                break
-
-    def end_streaming(self) -> None:
-        if self.__streams and not self.__streams.cancelled():
-            self.__streams.cancel()
+        super().__init__(
+            stub=TranscriptGrpcModule.TranscriptStub(channel),
+            request=TranscriptModule.TranscriptRequest(),
+            metadata=metadata,
+        )
